pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "image-publisher"
version = "0.1.0"
description = "A Python utility to read images from a directory and publish them to a Solace message broker."
readme = "README.md"
requires-python = ">=3.6"
dependencies = [
    "solace-pubsubplus>=1.3.0"
]

[project.scripts]
image-publisher = "image_publisher:main"
